+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Changed
 - 
+- Logging in `Packer` class to utilize the standart new in `et`.
 
 ### Fixed
 - 
@@ -38,7 +39,6 @@
 ### Fixed
 - Mistakes in `setup.py` `main` function code parts for creating a new project's `main.py` and `paths.py` to actually use the user provided software name.
 - Removed author from `pyproject.toml` file.
-
 - Git branch deletion, to first switch to the main branch, since it isn't possible to delete the branch currently in use.
 - Swapped the place of the code responsible for writing `CHANGELOG.md` new version's boilerplate code to happen after the branch, since otherwise git would discard changes.
 - Merging logic to just always prefer "theirs" changes or the development branches version in case of conflict.